If CUPS takes a while to respond a request, and in the meanwhile the user closes the application, our worker thread will still be running and when the CUPS request finishes the waitForFinish() event loop exits continuing to do a change on the GUI which was closed but not deleted.
The application then crashes inside Qt because we tried to set something
on a widget that was not visible or so.

Since waitForFinish() is not very cool and must be avoided in this case
so we connect all the affected requests to a slot that checks if the
request fails, and then requests the an UI update to use the values
from the model.

BUG: 319889