Initialize the translations of the ppd
Summary:
The to ppdLocalize() make sure that the translated strings in the PPD for
the current language are returned instead of the untranslated string.

Unfortunately not all translations are returned, for example section
titles are not translated (the translations are defined a bit differently
in the PPD file and I didn't fine an easy way in the API to extract them).

BUG: 387494

Test Plan:
(Most) translations are now available for the printer reported in the bug,
and at least another printer.

Reviewers: dantti, aacid

Reviewed By: aacid

Subscribers: ngraham, kde-utils-devel, #localization

Differential Revision: https://phabricator.kde.org/D9996