Don't fetch server settings for preferences button
Summary:
When the user is not member of a system group pressing the System Preferences button triggers a password dialog because of the getServerSettings() call. When the dialog finished the menu is closed again. We don't actually need to fetch the settings since its already done on startup

Needs D23403 to be effective.

Reviewers: dantti, broulik

Reviewed By: dantti

Subscribers: kde-utils-devel

Differential Revision: https://phabricator.kde.org/D23407