[Applet] Fix list item margins
Summary:
BUG: 360199
BUG: 398681
FIXED-IN: 20.04.0

This is a minimal change needed to fix the aforementioned bugs. Ultimately I would like to
make a re-usable component to implement this "expandable list item" paradigm (which is
also used in the Networks, Bluetooth, Vaults, and Device Notifier applets) so we don't
have to maintain five versions of the code. See T12812.

Test Plan:
Before: {F8171128}

After: {F8171127}

Reviewers: #plasma, dantti, cblack, ndavis

Reviewed By: cblack, ndavis

Subscribers: cblack, sefaeyeoglu, kde-utils-devel

Differential Revision: https://phabricator.kde.org/D27995