[applet] Simplify printer list item subtitle
The subtitle currently builds up a complicated string that tells you
thhe number of active and/or inactive jobs. However what you really
care about is the current state: idle or printing, paused, stuck, etc.
I don't think The number of jobs is useful or actionable information.
So let's remove it.

This happens to fix a bug whereby the state is displayed incorrectly.

BUG: 425534
FIXED-IN: 20.12.0