[applet] Remove config dialog and put options in Hamburger menu
This commit deletes the applet's config dialog and puts its three
mutually-exclusive options for what to show into the applet's hamburger
menu, exactly like how the Disks & Devices applet does it. This allows
removing the action to show the applet's config dialog, which means
there is now only a single "configure" action and it opens the KCM. The
net result is a UI that is cleaner, less confusing, and more consistent,
with the options made more discoverable.

BUG: 407857
FIXED-IN: 21.08