Set libkcups VERSION and SOVERSION
It's a private library so it doesn't need to maintain a SOVERSION
that matches the ABI, but it is installed in the default loader
directory so it should get a version that changes when the ABI changes.

LIBRARY NAMELINK_SKIP in install() makes it so that the unversioned
symlink isn't installed.