kcm: Handle recommended drivers not found scenario
For discovered devices, we query open systems printing for
recommended drivers.  It could be that there are no recommended
drivers found or the service is not installed, so give the ability
to select the driver manually as a fallback.

BUG: 482409
(cherry picked from commit ab7c5e31debfd179209a85ae442000f3a2b0a57f)