kcm: Provide System-Config-Printer package status
Add a method to the kcm to query SCP installed status. Also add optional
cmake install for SCP if PackageKit is available.  If PackageKit is not
available, the option can be disabled.

Add banner to main kcm view notifying user of PackageKit and/or SCP
installed status, with an action to install SCP.