libkcups: KCupsRequest: Add new apis and signals
Use cupsEnumDests to get printers.  This is the recommended api
for moving away from the deprecated CUPS apis that will be removed
in CUPS 3.x.

Use cupsCopyDestInfo to get PPD files. This is the recommended api for
CUPS 3.x for moving from deprecated CUPS api getPPD2.

Also, add an method to get printer attributes via a signal
instead of loading a list.  This can then be used by models to add
or update model items.