libkcups: Refactor PrinterModel to use new KCupsRequest apis
Currently the printer model tells the request to get printers via
and IPP request, which then populates a list and notifies that the list
is there.  Then, the model copies that list from the request into
the model.

This refactor uses the new KCupsRequest::getDestination to get printer
devices.  For each device found, the callback then signals the device map
to the model for insert/update.

Also, all CUPS requests are created with the global CUPS connection.  This has
fixed a bug with CUPS notify subscriptions (via Dbus) and allows us to
remove the deprecated Redhat dbus signals for CUPS server/device
changes.