Require CUPS version 2.4.x
CUPS 2.4.x provides the newer IPP Everywhere features but does not
remove the deprecated items that actually will be removed with v3.x.

This allows us to take some small steps preparing for some of the 3.x
changes while still providing the basics that are in place now and
depend on deprecated features.