kcm/[core,qml]: Add signals/handlers for the kcm command line, qml
When calling the kcm with additional args (--args) or when an
activation request (with args) occurs, parse the commandline and
signal for the appropriate command.

Update main.qml to handle the command line signals.

This allows the existing kded (and other processes) to use the kcm for printer
"configure" events.