RequestActivate on the printer manager helper dialog windows
If a corresponding dialog for a given printer (e.g. print queue) is already open,
call requestActivate on it to ensure it comes to the front.

CommandLauncherApp in ProcessRunner already sets up XDG Activation, and Qt
will use it in requestActivate but the app still has to call it.


(cherry picked from commit 243a9f9d1368710b5bdaf3d39fbdf19caa4b5f60)

Co-authored-by: Kai Uwe Broulik <kde@privat.broulik.de>